+++ CHANGELOG.md
@@ -41,18 +41,19 @@
 ### TODO
 
 * [open issues](https://github.com/Vanessa219/vditor/issues)
+* [119](https://github.com/Vanessa219/vditor/issues/119) 所见即所得模式支持脚注 `引入特性`
 
 ### v2.2.19 / 未发布
 
+* [206](https://github.com/Vanessa219/vditor/issues/206) combined blockquote and lists (windows firefox) `修复缺陷`
+* [151](https://github.com/Vanessa219/vditor/issues/151) 支持隐藏编辑器工具栏 `引入特性`
 * [121](https://github.com/Vanessa219/vditor/issues/121) 所见即所得模式支持 [ToC] `引入特性`
-* [119](https://github.com/Vanessa219/vditor/issues/119) 所见即所得模式支持脚注 `引入特性`
 * [55](https://github.com/Vanessa219/vditor/issues/55) 所见即所得模式支持链接引用定义`引入特性`
 
 ### v2.2.18 / 2020-03-05
 
 * [209](https://github.com/Vanessa219/vditor/pull/209) 支持 Graphviz `引入特性`
 * [207](https://github.com/Vanessa219/vditor/issues/207) heading at first time(windows firefox) `修复缺陷`
-* [206](https://github.com/Vanessa219/vditor/issues/206) combined blockquote and lists (windows firefox) `修复缺陷`
 * [205](https://github.com/Vanessa219/vditor/issues/205) not working typewrite mode (windows firefox) `修复缺陷`
 * [204](https://github.com/Vanessa219/vditor/issues/204) Empty Enter in Firefox `修复缺陷`
 * [203](https://github.com/Vanessa219/vditor/issues/203) when cache:false, use cache `修复缺陷`
@@ -126,6 +127,7 @@
   * 移除 index-preview.html, index-preview.js 文件
   * 添加 graphvizRender
   * 添加 option.preview.markdown.toc
+  * 添加 option.hideToolbar
 
 ### v2.1.15 / 2020-02-09