+++ CHANGELOG.md
@@ -49,6 +49,10 @@
 
 * [open issues](https://github.com/Vanessa219/vditor/issues)
 
+### v2.3.0 / 未发布
+
+* [216](https://github.com/Vanessa219/vditor/issues/216) subtoolbar 向下溢出 `修复缺陷`
+
 ### v2.2.19 / 2020-03-10
 
 * [214](https://github.com/Vanessa219/vditor/issues/214) wysiwyg heading id `引入特性`